requirements.txt
requests
python-dotenv
simpleeval